requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
gunicorn==20.1.0
celery==5.2.7
redis==4.5.5
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==9.5.0
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-health-check==3.17.0
factory-boy==3.2.1
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.5
django-debug-toolbar==4.1.0
